feat: refact witness
