+++ algebraic/src/circom_witness/memory.rs
