+++ plonky/src/lib.rs
@@ -12,10 +12,10 @@ extern crate num_traits;
 extern crate rand;
 
 pub use algebraic::circom_circuit;
+pub use algebraic::circom_witness;
 pub use algebraic::errors;
 pub use algebraic::field_gl;
 pub use algebraic::r1cs;
-pub use algebraic::r1cs_witness;
 
 pub mod plonk;
 pub mod reader;
