+++ plonky/src/reader.rs
@@ -94,7 +94,7 @@ pub fn maybe_load_key_lagrange_form<E: Engine>(
     }
 }
 
-/// load r1cs_witness file by filename with autodetect encoding (bin or json).
+/// load circom_witness file by filename with autodetect encoding (bin or json).
 pub fn load_witness_from_file<E: ScalarEngine>(filename: &str) -> Vec<E::Fr> {
     let file = OpenOptions::new()
         .read(true)
@@ -105,11 +105,11 @@ pub fn load_witness_from_file<E: ScalarEngine>(filename: &str) -> Vec<E::Fr> {
         load_witness_from_json::<E, BufReader<File>>(render)
     } else {
         load_witness_from_bin_reader::<E, BufReader<File>>(render)
-            .expect("read r1cs_witness failed")
+            .expect("read circom_witness failed")
     }
 }
 
-/// load r1cs_witness from json by a reader
+/// load circom_witness from json by a reader
 fn load_witness_from_json<E: ScalarEngine, R: Read>(reader: R) -> Vec<E::Fr> {
     let witness: Vec<String> = serde_json::from_reader(reader).expect("unable to read.");
     witness
@@ -118,12 +118,12 @@ fn load_witness_from_json<E: ScalarEngine, R: Read>(reader: R) -> Vec<E::Fr> {
         .collect::<Vec<E::Fr>>()
 }
 
-/// load r1cs_witness from u8 array
+/// load circom_witness from u8 array
 pub fn load_witness_from_array<E: ScalarEngine>(buffer: Vec<u8>) -> Result<Vec<E::Fr>> {
     load_witness_from_bin_reader::<E, _>(buffer.as_slice())
 }
 
-/// load r1cs_witness from u8 array by a reader
+/// load circom_witness from u8 array by a reader
 pub fn load_witness_from_bin_reader<E: ScalarEngine, R: Read>(mut reader: R) -> Result<Vec<E::Fr>> {
     let mut wtns_header = [0u8; 4];
     reader.read_exact(&mut wtns_header)?;
@@ -159,7 +159,7 @@ pub fn load_witness_from_bin_reader<E: ScalarEngine, R: Read>(mut reader: R) -> 
         return Err(EigenError::from("invalid curve prime".to_string()));
     }
     let witness_len = reader.read_u32::<LittleEndian>()?;
-    log::debug!("r1cs_witness len {}", witness_len);
+    log::debug!("circom_witness len {}", witness_len);
     let sec_type = reader.read_u32::<LittleEndian>()?;
     if sec_type != 2 {
         return Err(EigenError::from("invalid section type".to_string()));
@@ -167,7 +167,7 @@ pub fn load_witness_from_bin_reader<E: ScalarEngine, R: Read>(mut reader: R) -> 
     let sec_size = reader.read_u64::<LittleEndian>()?;
     if sec_size != (witness_len * field_size) as u64 {
         return Err(EigenError::from(format!(
-            "Invalid r1cs_witness section size {}",
+            "Invalid circom_witness section size {}",
             sec_size
         )));
     }
